Make candlepin::repo a standalone class
